# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Main API",
    "version": "1.0.0",
    "description": "API для работы с пользователями, постами и комментариями"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "сервер"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sessionId"
      }
    },
    "schemas": {
...
      "ErrorUnauthorized": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Unauthorized"
          },
          "error": {
            "type": "string",
            "example": "Invalid token"
          },
          "statusCode": {
            "type": "integer",
            "example": 401
          }
        }
      },
      "ErrorForbidden": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Forbidden"
          },
          "error": {
            "type": "string",
            "example": "You do not have access"
          },
          "statusCode": {
            "type": "integer",
            "example": 403
          }
        }
      },
      "ErrorValidation": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Validation failed"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "statusCode": {
            "type": "integer",
            "example": 400
          }
        }
      },
      "ErrorNotFound": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Not found"
          },
          "error": {
            "type": "string",
            "example": "Resource not found"
          },
          "statusCode": {
            "type": "integer",
            "example": 404
          }
        }
      },
...
      "Comment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "clnjak7xj000008l0a9zq3k4f"
          },
          "content": {
            "type": "string",
            "example": "Хорошая статья!"
          },
          "authorId": {
            "type": "string",
            "example": "clnja0001user123456789"
          },
          "postId": {
            "type": "string",
            "example": "clp123456post987654321"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2023-10-01T12:00:00Z"
          }
        }
      },
      "CommentCreate": {
        "type": "object",
        "required": [
          "content"
        ],
        "properties": {
          "content": {
            "type": "string",
            "minLength": 2,
            "maxLength": 1000,
            "example": "Очень полезный пост!"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/posts/{postId}/comments": {
      "post": {
        "summary": "Создать новый комментарий",
        "description": "Добавляет комментарий к указанному посту от имени авторизованного пользователя",
        "tags": [
          "Comments"
        ],
        "operationId": "createComment",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "postId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Идентификатор поста, к которому добавляется комментарий"
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Текст комментария",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Комментарий успешно создан",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comment"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка валидации входных данных",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorValidation"
                }
              }
            }
          },
          "401": {
            "description": "Неавторизованный доступ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorUnauthorized"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Получить комментарии к посту",
        "description": "Возвращает список всех комментариев, связанных с указанным постом",
        "tags": [
          "Comments"
        ],
        "operationId": "getPostComments",
        "parameters": [
          {
            "in": "path",
            "name": "postId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Идентификатор поста"
          }
        ],
        "responses": {
          "200": {
            "description": "Список комментариев",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Comment"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Пост не найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorNotFound"
                }
              }
            }
          }
        }
      }
    },
    "/api/comments/{id}": {
      "delete": {
        "summary": "Удалить комментарий",
        "description": "Удаляет комментарий по его ID, если у пользователя есть на это права",
        "tags": [
          "Comments"
        ],
        "operationId": "deleteComment",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Идентификатор комментария"
          }
        ],
        "responses": {
          "204": {
            "description": "Комментарий успешно удален",
            "content": {}
          },
          "401": {
            "description": "Неавторизованный доступ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorUnauthorized"
                }
              }
            }
          },
          "403": {
            "description": "Нет прав на удаление",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorForbidden"
                }
              }
            }
          },
          "404": {
            "description": "Комментарий не найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorNotFound"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}